Move smart contracts into separate module
