+++ iroha/src/smartcontracts.rs
@@ -1,0 +1,7 @@
+//! Module that opts different smartcontract runtime
+//!
+//! Currently supported only Iroha instructions
+
+pub mod isi;
+
+pub use isi::*;
