+++ iroha/src/smartcontracts/isi/query.rs
