+++ iroha/src/wsv.rs
@@ -12,7 +12,7 @@ use iroha_data_model::{domain::DomainsMap, peer::PeersIds, prelude::*};
 use iroha_error::Result;
 use tokio::task;
 
-use crate::{block::Chain, isi::FindError, prelude::*};
+use crate::{block::Chain, prelude::*, smartcontracts::FindError};
 
 /// Current state of the blockchain alligned with `Iroha` module.
 #[derive(Debug, Clone)]
